Move Version Requests to dataplane
