+++ src/dataplane-protocol/src/apis.rs
@@ -4,8 +4,7 @@
 //! Stores Api Keys supported by the SC.
 //!
 
-use dataplane::derive::Encode;
-use dataplane::derive::Decode;
+use crate::derive::{Decode, Encode};
 
 /// API call from client to SPU
 #[fluvio(encode_discriminant)]
@@ -13,7 +12,6 @@ use dataplane::derive::Decode;
 #[repr(u16)]
 pub enum AdminPublicApiKey {
     ApiVersion = 18,
-
     Create = 1001,
     Delete = 1002,
     List = 1003,
