+++ src/sc-schema/src/lib.rs
@@ -1,4 +1,3 @@
-mod apis;
 pub mod topic;
 pub mod spu;
 pub mod spg;
@@ -8,8 +7,6 @@ pub mod objects;
 mod request;
 mod response;
 
-use thiserror::Error;
-pub use apis::*;
 pub use request::*;
 pub use response::*;
 pub use admin::*;
@@ -18,16 +15,12 @@ pub mod errors {
     pub use dataplane::ErrorCode;
 }
 
-pub mod core {
-    pub use fluvio_controlplane_metadata::core::*;
-}
-
-pub mod store {
-    pub use fluvio_controlplane_metadata::store::*;
-}
+pub use dataplane::apis::AdminPublicApiKey;
+pub use fluvio_controlplane_metadata::core as core;
+pub use fluvio_controlplane_metadata::store as store;
 
 /// Error from api call
-#[derive(Error, Debug)]
+#[derive(thiserror::Error, Debug)]
 pub enum ApiError {
     #[error("Received error code: {0:#?} ({1:?})")]
     Code(dataplane::ErrorCode, Option<String>),
@@ -36,7 +29,6 @@ pub enum ApiError {
 }
 
 mod admin {
-
     use dataplane::api::Request;
 
     pub trait AdminRequest: Request {}
