+++ src/sc-schema/src/versions.rs
@@ -3,17 +3,13 @@
 //!
 //! # API Versions
 //!
-//! Public API to retrive a list of APIs and their version numbers from the SC.
+//! Public API to retrieve a list of APIs and their version numbers from the SC.
 //! SC supports Kafka as well as Fluvio specific APIs.
 //!
 
-use dataplane::api::Request;
-use dataplane::derive::{Decode, Encode};
-use dataplane::ErrorCode;
-
+pub use dataplane::versions::{ApiVersionKey, ApiVersions};
+pub use dataplane::versions::{ApiVersionsRequest, ApiVersionsResponse};
 use crate::AdminPublicApiKey;
-
-pub type ApiVersions = Vec<ApiVersionKey>;
 
 /// Given an API key, it returns max_version. None if not found
 pub fn lookup_version(api_key: AdminPublicApiKey, versions: &[ApiVersionKey]) -> Option<i16> {
@@ -23,33 +19,4 @@ pub fn lookup_version(api_key: AdminPublicApiKey, versions: &[ApiVersionKey]) ->
         }
     }
     None
-}
-
-// -----------------------------------
-// ApiVersionsRequest
-// -----------------------------------
-
-#[derive(Decode, Encode, Default, Debug)]
-pub struct ApiVersionsRequest {}
-
-impl Request for ApiVersionsRequest {
-    const API_KEY: u16 = AdminPublicApiKey::ApiVersion as u16;
-    type Response = ApiVersionsResponse;
-}
-
-// -----------------------------------
-// ApiVersionsResponse
-// -----------------------------------
-
-#[derive(Decode, Encode, Default, Debug)]
-pub struct ApiVersionsResponse {
-    pub error_code: ErrorCode,
-    pub api_keys: Vec<ApiVersionKey>,
-}
-
-#[derive(Decode, Encode, Default, Debug)]
-pub struct ApiVersionKey {
-    pub api_key: i16,
-    pub min_version: i16,
-    pub max_version: i16,
 }
